Add a script to update the latest Git version
This re-implements the `update-git-version` script as Ruby script: It
does not make much sense to use too many different languages in the same
project unless there are compelling reasons. We already use Ruby in this
project, and there is no compelling reason to add node.js to the fray.
Therefore a Ruby script it is.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>