docs(zh_HANS-CN): a trailing "full stop" character is not part of a URL
When encountering a bare URL, AsciiDoc helpfully renders it as a link.
In English, when encountering a bare URL at the end of a sentence,
AsciiDoc is smart enough not to include the full stop character (`.`)
that comes after the URL.

In other languages, other full stop characters exist, and AsciiDoc
apparently does not know the special meaning of `。` mistaking it for
being part of the bare URL. Let's work around that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>