redirect_to: keep the anchor, if any was specified
This is important because some translated manual pages point to
non-existing translated pages with an anchor. For example, the German
version of the `git` manual page points to
`docs/user-manual/de#git-concepts`, i.e. to the `git-concepts` anchor of
the German version of the `user-manual` (that does not exist yet).

Let's make sure that we not only redirect to an existing page (read: the
English one) in that case, but also jump to the specified anchor.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>